backend/requirements.txt
# Agents & Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# APIs
tavily-python>=0.5.0

# Config & Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0

# Paid.ai — Facturation agentique
paid-python>=1.0.0
